+++ examples/test_mod.py
@@ -5,18 +5,20 @@
 # exe = "notepad.exe"
 # start_paused = false
 # start_exe = false
+# interactive_console = false
 #
 # [tool.pymhf.gui]
 # always_on_top = false
 # 
 # [tool.pymhf.logging]
+# shown = false
 # log_dir = "."
 # log_level = "info"
 # window_name_override = "pyMHF Test Mod"
 # ///
 from logging import getLogger
 
-from pymhf import Mod
+from pymhf import Mod, load_mod_file
 
 logger = getLogger("TestMod")
 
@@ -28,3 +30,7 @@ class TestMod(Mod):
     def __init__(self):
         super().__init__()
         logger.info("Loaded Test mod!")
+
+
+if __name__ == "__main__":
+    load_mod_file(__file__)