+++ pymhf/utils/parse_toml.py
@@ -1,7 +1,9 @@
 import re
-from typing import Optional
+from typing import Optional, cast
 
 import tomlkit
+
+from pymhf.core._types import pymhfConfig
 
 REGEX = r"(?m)^# /// (?P<type>[a-zA-Z0-9-]+)$\s(?P<content>(^#(| .*)$\s)+)^# ///$"
 
@@ -25,17 +27,18 @@ def read_inline_metadata(script: str) -> Optional[tomlkit.TOMLDocument]:
         return None
 
 
-def _parse_toml(fpath: str, standalone: bool = False) -> Optional[dict]:
+def _parse_toml(fpath: str, standalone: bool = False) -> Optional[pymhfConfig]:
     settings = {}
     with open(fpath, "r") as f:
         if standalone:
             settings = read_inline_metadata(f.read())
         else:
             settings = tomlkit.loads(f.read())
-    return settings
+    if settings:
+        return cast(pymhfConfig, settings.value)
 
 
-def read_pymhf_settings(fpath: str, standalone: bool = False) -> dict:
+def read_pymhf_settings(fpath: str, standalone: bool = False) -> pymhfConfig:
     settings = _parse_toml(fpath, standalone)
     if not settings:
         return {}